requirements.txt
flask>=1.0.3
flask_bootstrap>=3.3.7.1
pyyaml>=5.1.1
nltk>=3.4.3